+++ vlinder/tests/pages/homepage.test.js
@@ -1,14 +1,12 @@
 import React from 'react';
 import { render, screen, fireEvent, waitFor } from '@testing-library/react';
 import HomePage from '@/app/homepage/page';
-import { createClient } from '@/utils/supabase/client';
+import { mockProfiles } from '@/__mocks__/supabaseClientMock';
 
-
-
-jest.mock('@/utils/supabase/client', () => ({
-  createClient: jest.fn(() => ({
+jest.mock('@/utils/supabase/client', () => {
+  const createClientMock = () => ({
     auth: {
-      getUser: jest.fn().mockResolvedValue({ data: { user: {id: 'test-user-id'} } })
+      getUser: jest.fn().mockResolvedValue({ data: { user: { id: 'test-user-id' } } })
     },
     from: jest.fn((table) => ({
       select: jest.fn().mockReturnThis(),
@@ -19,74 +17,20 @@ jest.mock('@/utils/supabase/client', () => ({
         if (table === 'profiles') cb({ data: mockProfiles });
       })
     }))
-  }))
-}));
+  });
+  return { createClient: createClientMock };
+});
 
 jest.mock('next/navigation', () => ({
   useRouter: jest.fn().mockReturnValue({
     push: jest.fn(),
   }),
-  }));
+}));
 
-  const mockProfiles = [
-    {
-      id: '1',
-      username: 'dragon',
-      avatar_url: null,
-      birthday: '1990-01-01',
-      gender: 'Male',
-      smoker: false,
-      sexual_orientation: 'heterosexual',
-      sex_positive: true,
-      display_disability: true,
-      disability: ['vision'],
-      description: 'Profile description',
-      need_assistance: false,
-      profile_hobbies: [
-        { hobbies: { id: 1, name: 'Reading', emoji: '📚' } },
-        { hobbies: { id: 2, name: 'Gaming', emoji: '🎮' } }
-      ]
-    },
-    {
-      id: '2',
-      username: 'unicorn',
-      avatar_url: null,
-      birthday: '1995-05-15',
-      gender: 'Other',
-      smoker: true,
-      sexual_orientation: 'asexual',
-      sex_positive: false,
-      display_disability: false,
-      disability: [],
-      description: 'Profile description',
-      need_assistance: true,
-      profile_hobbies: [
-        { hobbies: { id: 3, name: 'Cooking', emoji: '🍳' } },
-        { hobbies: { id: 4, name: 'Hiking', emoji: '🥾' } }
-      ]
-    },
-    {
-      id: '3',
-      username: 'mamaBear',
-      avatar_url: null,
-      birthday: '1985-10-10',
-      gender: 'Female',
-      smoker: false,
-      sexual_orientation: 'heterosexual',
-      sex_positive: true,
-      display_disability: true,
-      disability: ['hearing'],
-      description: 'Profile description',
-      need_assistance: false,
-      profile_hobbies: [
-        { hobbies: { id: 5, name: 'Swimming', emoji: '🏊' } },
-        { hobbies: { id: 6, name: 'Traveling', emoji: '✈️' } }
-      ]
-    }
-  ];
-
-  describe('HomePage', () => {
-
+describe('HomePage', () => {
+  beforeEach(() => {
+    jest.clearAllMocks();
+  });
 
     it('renders loading state initially with skeleton cards', () => {
       render(<HomePage />);
