+++ vlinder/tsconfig.json
@@ -25,6 +25,6 @@
 
     }
   },
-  "include": ["next-env.d.ts", "**/*.ts", "**/*.tsx", ".next/types/**/*.ts"],
+  "include": ["next-env.d.ts", "**/*.ts", "**/*.tsx", ".next/types/**/*.ts", "__mocks__/next/navigation.js", "__mocks__/supabase/client.js"],
   "exclude": ["node_modules"]
 }
